+++ MinimOsd_Extra/compat.h
@@ -28,4 +28,9 @@ void WEAK __cxa_guard_abort (__guard *) {};
 void WEAK __cxa_pure_virtual(void) {};
 */
 
+template<typename T> void ZeroIt(T& value)
+{
+    memset(&value,0,sizeof(value));
+}
+
 #endif